+++ lib/src/upgrade_card.dart
@@ -13,10 +13,20 @@ class UpgradeCard extends UpgradeBase {
   /// `EdgeInsets.all(4.0)`.
   final EdgeInsetsGeometry margin;
 
+  /// An optional maximum number of lines for the text to span, wrapping if necessary.
+  final int? maxLines;
+
+  /// How visual overflow should be handled.
+  final TextOverflow? overflow;
+
   /// Creates a new [UpgradeCard].
-  UpgradeCard(
-      {Key? key, Upgrader? upgrader, this.margin = const EdgeInsets.all(4.0)})
-      : super(upgrader ?? Upgrader.sharedInstance, key: key);
+  UpgradeCard({
+    super.key,
+    Upgrader? upgrader,
+    this.margin = const EdgeInsets.all(4.0),
+    this.maxLines = 15,
+    this.overflow = TextOverflow.ellipsis,
+  }) : super(upgrader ?? Upgrader.sharedInstance);
 
   /// Describes the part of the user interface represented by this widget.
   @override
@@ -48,8 +58,9 @@ class UpgradeCard extends UpgradeBase {
 
   /// Build the UpgradeCard Widget.
   Widget buildUpgradeCard(BuildContext context, UpgradeBaseState state) {
-    final title = upgrader.messages.message(UpgraderMessage.title);
-    final message = upgrader.message();
+    final appMessages = upgrader.determineMessages(context);
+    final title = appMessages.message(UpgraderMessage.title);
+    final message = upgrader.body(appMessages);
     final releaseNotes = upgrader.releaseNotes;
     final shouldDisplayReleaseNotes = upgrader.shouldDisplayReleaseNotes();
     if (upgrader.debugLogging) {
@@ -70,14 +81,12 @@ class UpgradeCard extends UpgradeBase {
             mainAxisSize: MainAxisSize.min,
             crossAxisAlignment: CrossAxisAlignment.start,
             children: <Widget>[
-              Text(
-                  Upgrader().messages.message(UpgraderMessage.releaseNotes) ??
-                      '',
+              Text(appMessages.message(UpgraderMessage.releaseNotes) ?? '',
                   style: const TextStyle(fontWeight: FontWeight.bold)),
               Text(
                 releaseNotes,
-                maxLines: 15,
-                overflow: TextOverflow.ellipsis,
+                maxLines: maxLines,
+                overflow: overflow,
               ),
             ],
           ));
@@ -97,15 +106,14 @@ class UpgradeCard extends UpgradeBase {
                 Padding(
                     padding: const EdgeInsets.only(top: 15.0),
                     child: Text(
-                        upgrader.messages.message(UpgraderMessage.prompt) ??
-                            '')),
+                        appMessages.message(UpgraderMessage.prompt) ?? '')),
                 if (notes != null) notes,
               ],
             ),
             actions: <Widget>[
               if (upgrader.showIgnore)
                 TextButton(
-                    child: Text(upgrader.messages
+                    child: Text(appMessages
                             .message(UpgraderMessage.buttonTitleIgnore) ??
                         ''),
                     onPressed: () {
@@ -117,9 +125,9 @@ class UpgradeCard extends UpgradeBase {
                     }),
               if (upgrader.showLater)
                 TextButton(
-                    child: Text(upgrader.messages
-                            .message(UpgraderMessage.buttonTitleLater) ??
-                        ''),
+                    child: Text(
+                        appMessages.message(UpgraderMessage.buttonTitleLater) ??
+                            ''),
                     onPressed: () {
                       // Save the date/time as the last time alerted.
                       upgrader.saveLastAlerted();
@@ -128,9 +136,9 @@ class UpgradeCard extends UpgradeBase {
                       state.forceUpdateState();
                     }),
               TextButton(
-                  child: Text(upgrader.messages
-                          .message(UpgraderMessage.buttonTitleUpdate) ??
-                      ''),
+                  child: Text(
+                      appMessages.message(UpgraderMessage.buttonTitleUpdate) ??
+                          ''),
                   onPressed: () {
                     // Save the date/time as the last time alerted.
                     upgrader.saveLastAlerted();